refactor: extract shared utilities and convert to .js
- Create build-fixtures.utilities.js with discoverFixtures() and
  extractFTemplates() shared by both build scripts
- Convert build-fixtures-with-webui.mjs to .js (package uses ESM)
- Both scripts now import from the shared utilities module
