+++ packages/fast-html/package.json
@@ -37,7 +37,7 @@
     "test:playwright": "playwright test",
     "test:chromium": "playwright test --project=chromium",
     "test:rules": "sg test --skip-snapshot-tests",
-    "test:webui-integration": "node scripts/build-fixtures-with-webui.mjs",
+    "test:webui-integration": "node scripts/build-fixtures-with-webui.js",
     "test": "npm run test:playwright",
     "test:ui": "concurrently -k -n fast-element,fast-html,playwright \"npm run dev --workspace=@microsoft/fast-element\" \"npm:watch\" \"npx playwright test --ui\"",
     "watch": "npm run build:tsc -- -w --preserveWatchOutput"
