+++ packages/fast-html/scripts/build-fixtures.js
@@ -1,26 +1,14 @@
 #!/usr/bin/env node
 import { execFileSync } from "node:child_process";
-import { existsSync, readdirSync, readFileSync, writeFileSync } from "node:fs";
+import { readFileSync, writeFileSync } from "node:fs";
 import { createRequire } from "node:module";
 import { dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
+import { discoverFixtures } from "./build-fixtures.utilities.js";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const fixturesDir = resolve(__dirname, "../test/fixtures");
-
-// Auto-discover fixture directories that contain the required files.
-const fixtures = readdirSync(fixturesDir, { withFileTypes: true })
-    .filter(entry => {
-        if (!entry.isDirectory()) return false;
-        const dir = join(fixturesDir, entry.name);
-        return (
-            existsSync(join(dir, "entry.html")) &&
-            existsSync(join(dir, "templates.html")) &&
-            existsSync(join(dir, "state.json"))
-        );
-    })
-    .map(entry => entry.name)
-    .sort();
+const fixtures = discoverFixtures(fixturesDir);
 
 const require = createRequire(import.meta.url);
 const fastBin = require.resolve("@microsoft/fast-build/bin/fast.js");
